# src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "InvoiceReviewAssistant.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5190/"
    }
  ],
  "paths": {
...
    "/api/invoices/{id}/history": {
      "get": {
        "tags": [
          "InvoiceHistory"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditHistoryPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "getInvoiceHistory"
      }
    },
    "/api/invoices": {
      "get": {
        "tags": [
          "InvoiceQueue"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/InvoiceStatus"
              }
            },
            "style": "form",
            "explode": true
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/InvoiceSort",
              "default": "updatedAtDesc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceQueuePageDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "listInvoices"
      },
      "post": {
        "tags": [
          "InvoiceUploads"
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDto"
                }
              }
            },
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "uploadInvoice",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}/reject": {
      "post": {
        "tags": [
          "InvoiceRejections"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectInvoiceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceProblemDetails"
                }
              }
            },
            "headers": {
              "X-Correlation-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "operationId": "rejectInvoice"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuditHistoryPageDto": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalItems",
          "totalPages",
          "hasPreviousPage",
          "hasNextPage"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditEventDto"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "totalItems": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "DocumentTextSource": {
        "type": "string",
        "enum": [
          "nativeText",
          "ocr"
        ]
      },
...
      "InvoiceDetailDto": {
        "required": [
          "id",
          "status",
          "draftVersion",
          "lastValidatedVersion",
          "createdAt",
          "updatedAt",
          "document",
          "documentTextSource",
          "fields",
          "reviewNotes",
          "summary",
          "currentValidation",
          "corrections",
          "processingFailure",
          "decision"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "lastValidatedVersion": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32",
            "minimum": 1
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "document": {
            "$ref": "#/components/schemas/InvoiceDocumentDto"
          },
          "documentTextSource": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentTextSource"
              }
            ]
          },
          "fields": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InvoiceFieldsDto"
              }
            ]
          },
          "reviewNotes": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "$ref": "#/components/schemas/InvoiceSummaryDto"
          },
          "currentValidation": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ValidationRunDto"
              }
            ]
          },
          "corrections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldCorrectionDto"
            }
          },
          "processingFailure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProcessingFailureDto"
              }
            ]
          },
          "decision": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/InvoiceDecisionDto"
              }
            ]
          }
        },
        "additionalProperties": false
      },
      "InvoiceDocumentDto": {
        "required": [
          "originalFilename",
          "mediaType",
          "byteLength",
          "sha256",
          "pageCount",
          "integrityStatus"
        ],
        "type": "object",
        "properties": {
          "originalFilename": {
            "type": "string"
          },
          "mediaType": {
            "type": "string",
            "enum": [
              "application/pdf"
            ]
          },
          "byteLength": {
            "type": "integer",
            "format": "int64",
            "minimum": 1
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "pageCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "integrityStatus": {
            "$ref": "#/components/schemas/DocumentIntegrityStatus"
          }
        },
        "additionalProperties": false
      },
...
      "InvoiceFieldsDto": {
        "required": [
          "supplier",
          "reference",
          "datesAndTerms",
          "amounts"
        ],
        "type": "object",
        "properties": {
          "supplier": {
            "$ref": "#/components/schemas/SupplierFieldsDto"
          },
          "reference": {
            "$ref": "#/components/schemas/ReferenceFieldsDto"
          },
          "datesAndTerms": {
            "$ref": "#/components/schemas/DatesAndTermsFieldsDto"
          },
          "amounts": {
            "$ref": "#/components/schemas/AmountFieldsDto"
          }
        },
        "additionalProperties": false
      },
      "InvoiceProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 100
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          },
          "fields": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "object",
                "additionalProperties": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            ]
          },
          "currentVersion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "type": "integer",
                "format": "int32",
                "minimum": 1
              }
            ]
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "correlationId",
          "fields",
          "currentVersion"
        ],
        "additionalProperties": false
      },
      "InvoiceQueueItemDto": {
        "required": [
          "id",
          "status",
          "supplierName",
          "invoiceNumber",
          "invoiceDate",
          "total",
          "currency",
          "draftVersion",
          "warningCount",
          "errorCount",
          "exceptionCount",
          "processingFailure",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "supplierName": {
            "type": [
              "null",
              "string"
            ]
          },
          "invoiceNumber": {
            "type": [
              "null",
              "string"
            ]
          },
          "invoiceDate": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "total": {
            "type": [
              "null",
              "string"
            ],
            "pattern": "^-?(?:0|[1-9]\\d{0,26})\\.\\d{2}$"
          },
          "currency": {
            "type": [
              "null",
              "string"
            ]
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "warningCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "exceptionCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "processingFailure": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ProcessingFailureDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          }
        },
        "additionalProperties": false
      },
      "InvoiceQueuePageDto": {
        "required": [
          "items",
          "summary",
          "page",
          "pageSize",
          "totalItems",
          "totalPages",
          "hasPreviousPage",
          "hasNextPage"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceQueueItemDto"
            }
          },
          "summary": {
            "$ref": "#/components/schemas/InvoiceQueueSummaryDto"
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "totalItems": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "InvoiceQueueSummaryDto": {
        "required": [
          "totalInvoiceCount",
          "processingCount",
          "reviewRequiredCount",
          "readyForApprovalCount",
          "approvedCount",
          "rejectedCount",
          "processingFailedCount",
          "pendingReviewCount",
          "warningInvoiceCount",
          "errorInvoiceCount"
        ],
        "type": "object",
        "properties": {
          "totalInvoiceCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "processingCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "reviewRequiredCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "readyForApprovalCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "approvedCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "rejectedCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "processingFailedCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "pendingReviewCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "warningInvoiceCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorInvoiceCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "InvoiceSort": {
        "type": "string",
        "enum": [
          "updatedAtDesc",
          "updatedAtAsc",
          "createdAtDesc",
          "createdAtAsc"
        ]
      },
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "processing",
          "reviewRequired",
          "readyForApproval",
          "approved",
          "rejected",
          "processingFailed"
        ]
      },
      "InvoiceSummaryDto": {
        "required": [
          "extractedFieldCount",
          "warningCount",
          "errorCount",
          "manualCorrectionCount"
        ],
        "type": "object",
        "properties": {
          "extractedFieldCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "warningCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "errorCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "manualCorrectionCount": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
...
      "ProcessingFailureDto": {
        "required": [
          "stage",
          "code",
          "message",
          "failedAt"
        ],
        "type": "object",
        "properties": {
          "stage": {
            "$ref": "#/components/schemas/ProcessingStage"
          },
          "code": {
            "$ref": "#/components/schemas/ProcessingFailureCode"
          },
          "message": {
            "type": "string"
          },
          "failedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          }
        },
        "additionalProperties": false
      },
...
      "RejectInvoiceRequest": {
        "required": [
          "expectedVersion",
          "reason"
        ],
        "type": "object",
        "properties": {
          "expectedVersion": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "reason": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "AuditEventDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^(?:0|[1-9]\\d*)$"
          },
          "invoiceId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"
          },
          "type": {
            "$ref": "#/components/schemas/AuditEventType"
          },
          "actor": {
            "$ref": "#/components/schemas/AuditActor"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
          },
          "draftVersion": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "details": {}
        },
        "required": [
          "id",
          "invoiceId",
          "type",
          "actor",
          "occurredAt",
          "draftVersion",
          "details"
        ],
        "additionalProperties": false,
        "discriminator": {
          "propertyName": "type",
          "mapping": {
            "invoiceUploaded": "#/components/schemas/InvoiceUploadedAuditEventDto",
            "extractionCompleted": "#/components/schemas/ExtractionCompletedAuditEventDto",
            "extractionFailed": "#/components/schemas/ExtractionFailedAuditEventDto",
            "draftSaved": "#/components/schemas/DraftSavedAuditEventDto",
            "validationCompleted": "#/components/schemas/ValidationCompletedAuditEventDto",
            "invoiceApproved": "#/components/schemas/InvoiceApprovedAuditEventDto",
            "invoiceRejected": "#/components/schemas/InvoiceRejectedAuditEventDto",
            "documentIntegrityChanged": "#/components/schemas/DocumentIntegrityChangedAuditEventDto"
          }
        }
      }
    }
  },
...
}